refactor go app into packages
