+++ apps/go-server/main.go
@@ -2,13 +2,9 @@ package main
 
 import (
 	"embed"
-	"fmt"
-	"io/fs"
+	"goodwin/apps/go-server/handlers"
 	"log"
-	"net"
 	"net/http"
-	"os"
-	"path/filepath"
 	"time"
 
 	"github.com/gorilla/mux"
@@ -16,13 +12,23 @@ import (
 )
 
 //go:embed all:build/*
-var staticFiles embed.FS
+var files embed.FS
 var oscClient *osc.Client = osc.NewClient("localhost", 53000)
 
 func main() {
+	observers := make(map[string][]chan string)
 	router := mux.NewRouter()
-	router.Path(`/api/{rest:[a-zA-Z0-9=\-\/]+}`).Methods(http.MethodPost, http.MethodPut).HandlerFunc(handleOscCommand)
-	router.PathPrefix("/").HandlerFunc(handleStaticFiles)
+
+	router.
+		Path(`/api/{rest:[a-zA-Z0-9=\-\/]+}`).
+		Methods(http.MethodPost, http.MethodPut).
+		HandlerFunc(handlers.HandleOscCommand(oscClient, observers))
+
+	router.
+		PathPrefix("/").
+		Methods(http.MethodGet).
+		HandlerFunc(handlers.ServeStatic(files))
+
 	port := "5000"
 	srv := &http.Server{
 		Handler:      router,
@@ -32,96 +38,4 @@ func main() {
 	}
 	log.Println("listing on port", port)
 	log.Fatal(srv.ListenAndServe())
-}
-
-type OscCommand struct {
-	Arguments string
-}
-
-func handleOscCommand(w http.ResponseWriter, r *http.Request) {
-	// var command OscCommand
-
-	defer r.Body.Close()
-
-	params := mux.Vars(r)
-	address := "/" + params["rest"]
-	// decoder := json.NewDecoder(r.Body)
-
-	// if err := decoder.Decode(&command); err != nil {
-	// 	http.Error(w, err.Error(), http.StatusBadRequest)
-	// 	return
-	// }
-
-	msg := osc.NewMessage(address)
-	// if command.Arguments != "" {
-	// 	msg.Append(command.Arguments)
-	// }
-
-	oscClient.Send(msg)
-
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-}
-
-func handleStaticFiles(w http.ResponseWriter, r *http.Request) {
-	// get the absolute path to prevent directory traversal
-	path, err := filepath.Abs(r.URL.Path)
-	if err != nil {
-		// if we failed to get the absolute path respond with a 400 bad request and stop
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-	// prepend the path with the path to the static directory
-	path = filepath.Join("build", path)
-
-	_, err = staticFiles.Open(path)
-	if os.IsNotExist(err) {
-		// file does not exist, serve index.html
-		index, err := staticFiles.ReadFile(filepath.Join("build", "index.html"))
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		w.WriteHeader(http.StatusAccepted)
-		w.Write(index)
-		return
-	} else if err != nil {
-		// if we got an error (that wasn't that the file doesn't exist) stating the
-		// file, return a 500 internal server error and stop
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	// get the subdirectory of the static dir
-	statics, err := fs.Sub(staticFiles, "build")
-	// otherwise, use http.FileServer to serve the static dir
-	http.FileServer(http.FS(statics)).ServeHTTP(w, r)
-}
-
-func receiveFromQLab() {
-	conn, err := net.ListenUDP("udp", &net.UDPAddr{
-		Port: 53001,
-		IP:   net.ParseIP("127.0.0.1"),
-	})
-
-	defer conn.Close()
-
-	conn.SetReadBuffer(65_535)
-
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	for {
-		// TODO: review if necessary to read the maximum datagram size
-		buf := make([]byte, 65_535)
-		rlen, address, _ := conn.ReadFromUDP(buf[:])
-		fmt.Println(address)
-		go handleUDPPacket(buf, rlen)
-	}
-}
-
-func handleUDPPacket(buf []byte, rlen int) {
-	fmt.Println((buf[0:rlen]))
 }
